Delete image from delete function

+++ controllers/admin.js
@@ -186,9 +186,16 @@ exports.getProductsAll = (req, res) => {
 
 exports.postDeleteProduct = (req, res) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({
-			_id: prodId,
-			userId: req.user._id
+	Product.findById(prodId)
+		.then(product => {
+			if (!product) {
+				return next(new Error('Product not found'));
+			}
+			fileHelper.deleteFile(product.imageURL);
+			return Product.deleteOne({
+				_id: prodId,
+				userId: req.user._id
+			})
 		})
 		.then(result => {
 			res.redirect('/admin/products');
